Breeding changes and more pokemon
Before, breeding gave you a baby which was the same pokemon as the female, so if a charizard female was breeding, the baby would be a charizard, not a charmander.  This has supposedly been fixed through adding another parameter called evolutionTree.  Also, in adding the nidoran tree I have added a check for nidoran so that a female nidoran can breed either a female or male nidoran.
 I will test all of this when I'm done implementing all of the gen 1 pokemon.  Pokemon 1-34 have been implemented so far.