Saving implemented
FINALS ARE DONE, I'M BACK TO WORKING ON THIS
Saving is implemented but it is only local.  The plan is to eventually put this on a website so people can play with each other which means I will need to make saving work via a database but I'll make that when I finally make the website.