More Commenting and Fixed Issue with leveling
I originally made leveling without sorting in mind.  Now that sorting is in place, it didn't work in the case of the pc being sorted in level order.  To make it work, I added a getIndex() method to the Player class which takes an ownedPokemon as a parameter and returns its index in the pc.  Using that, I keep track of the pokemon's new position in the pc instead of always checking the original index the user input.  This should also act as a fix for alphabetical order sorting since pokemons' names change when they evolve and thus the pokemons' positions in the pc change.
The getIndex() method required that I make a .equals for OwnedPokemon and Pokemon.  The OwnedPokemon .equals makes sure EVERY data member of the OwnedPokemon passed as a parameter matches that of the OwnedPokemon it is being compared to.  The Pokemon .equals() just check for the given Pokemons' names to match since different Pokemon cannot have the same name.
I also noticed that level order sorting was ascending when it should have been descending so I fixed that.
As for commenting, I'm up to the evolve() method.