Gen 2
Finished Implementing Gen 2
Implemented all the special cases in breeding and evolution that came with the addition of gen 2
Made some changes to breeding.  Notably, any pokemon of the undiscovered egg group can breed with any pokemon of the undiscovered egg group.  If a player is breeding two pokemon of the undiscovered egg group, they get to choose which one the baby will be.
Players now get warned of how breeding works before they agree to do it.
Players can now choose not to breed in case they changed their mind or input the wrong pokemon