Started Gen 3 and Added Evolution Lock
I started adding gen 3 and reached a pokemon I prefer to remain unevolved when I train it, so I decided to add a Evolution Lock option so that the player can make it so their pokemon does not evolve as long as the lock is turned on.  The exception for this is special condition evolutions.  I assume that since the player chose to go to that option then they actually want to evolve it and its evolution lock shouldnt be considered.
I also redesigned the way the player is prompted about editing their pokemon when they are viewing their pokemon.  The prompts used to come in one after the other, but now the player is provided a menu with the available prompts so they can choose one.  After they choose one and complete the actions associated with that option, they are returned to the edit menu.  This keeps going until they choose to view another pokemon or to go back to the main menu.