Began Transitioning code to JavaScript
I have had the intention of making it so the game would run on a website.  All the java code I made was supposed to be the backend, with HTML and CSS as the frontend with JavaScript stitching everything together.
I'm having more trouble with it than I expected to.  I am currently working on learning JavaScript and React so that I can translate my java code to js and getting it working on a website (locally for now).