Creating OwnedPokemon and search()
Need to use Math.floor() instead of Math.round() when choosing a random number and using it as an array index.  Fixed issue of pokemon not having a nature which caused its stats to be calculated incorrectly.
Also made the spawnRateCounter array system which works just like in the main.java