Trying to work with React
I'm trying to work with React but keep running into issues.  I'll keep trying of course but at the rate things are going right now there might not be progress for a while.  Meanwhile I noticed that instead of using a makePokemon() function which returns an object with the given parameters, I can just directly make the object inside the array.  We'll see if I choose to do it that way instead, both seem to do the same thing more or less.
Also, the problem I'm facing right now is my search function.  It worked before, as can be seen from one of my earlier commits, but now that I'm using react, the page becomes unresponsive due to the while loop.  I'm really not sure why but I will of course try to fix it and progress further.